Sjy 1026 01

+++ ws/ForWorks/src/main/java/com/kh/forworks/notice/controller/NoticeController.java
@@ -17,6 +17,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import com.kh.forworks.FileUploader;
 import com.kh.forworks.PageVo;
 import com.kh.forworks.Pagination;
+import com.kh.forworks.department.vo.DepartmentVo;
+import com.kh.forworks.member.vo.MemberVo;
 import com.kh.forworks.notice.service.NoticeService;
 import com.kh.forworks.notice.vo.NoticeVo;
 import com.kh.forworks.noticeattachments.vo.NoticeAttachmentsVo;
@@ -34,10 +36,14 @@ public class NoticeController {
 
 	//공지사항 (화면)
 	@GetMapping("list/{pno}")
-	public String list(Model model,@PathVariable int pno, String keyword){
+	public String list(Model model,@PathVariable int pno, String keyword, HttpSession session){
 		
-		//System.out.println("key :: " +keyword);
-		
+		//로그인 연결x -> 로그인페이지로
+		MemberVo loginMember = (MemberVo)session.getAttribute("loginMember");
+		if (loginMember == null) {
+			session.setAttribute("toastMsg", "로그인이 필요합니다.");
+			return "member/login";
+		}else{
 		
 		//공지사항 전체 갯수
 		int totalCount = nts.selectTotalCount(keyword);
@@ -57,16 +63,18 @@ public class NoticeController {
 		}else {
 			return "error";
 		}
-		
+		}
 	}
 	
 	//공지사항 작성(화면)
 	@GetMapping("write")
-	public String write() {
+	public String write(Model model) {
 		
 		//부서목록 가져오기
+		List<DepartmentVo> dpvo = nts.selectAlldp();
+		//System.out.println(dpvo);
 		
-		
+		model.addAttribute("dpvo", dpvo);
 		return"notice/write";
 	}
 	
@@ -75,8 +83,10 @@ public class NoticeController {
 	public String write(NoticeVo ntvo, NoticeAttachmentsVo ntatVo ,HttpServletRequest req,Model model ,HttpSession session) {
 		System.out.println("공지::"+ntvo);
 		
-		ntvo.setEmpNo("1");
 		
+        MemberVo loginMember = (MemberVo)session.getAttribute("loginMember");
+        
+        ntvo.setEmpNo(loginMember.getEmpNo());
 		//파일 유무 확인
 		if (ntvo.getNtFileName() != null && !ntvo.getNtFileName().isEmpty()) {
 			//파일 있음
@@ -112,6 +122,8 @@ public class NoticeController {
 	@GetMapping("detail/{no}")
 	public String detail(@PathVariable String no, Model model) {
 		
+		//로그인한 회원이 접근권한이 있는지 판단[작성자의 부서 비교 게시글의 접근권한]
+		
 		//db 공지사항 조회
 		NoticeVo ntvo = nts.selectOne(no);
 		
@@ -129,15 +141,22 @@ public class NoticeController {
 	//공지사항 수정(화면)
 	@GetMapping("update/{no}")
 	public String update(@PathVariable String no, Model model) {
+		
+		//부서목록 가져오기
+		List<DepartmentVo> dpvo = nts.selectAlldp();
+		//System.out.println(dpvo);
+		
 		//db 공지사항 조회
 		NoticeVo ntvo = nts.selectOne(no);
 		
 		//첨부파일 확인
 		NoticeAttachmentsVo ntatVo = nts.checkFile(no);
 		System.out.println("파일 확인::"+ntatVo);
+		System.out.println(ntvo);
 		
 		model.addAttribute("ntvo", ntvo);
 		model.addAttribute("ntatVo", ntatVo);
+		model.addAttribute("dpvo",dpvo);
 		return"notice/update";
 	}
 	
@@ -157,16 +176,20 @@ public class NoticeController {
 		
 		//첨부파일 확인
 		ntatVo = nts.checkFile(no);
-		//System.out.println("파일 확인(정보수정post)::"+ntatVo);
+		System.out.println("파일 확인(정보수정post)::"+ntatVo);
 		
-		String fileName = ntatVo.getNtatOrigin();
-		File f =  new File(savePath +  fileName);
-		if(f.exists()) { //파일 존재하는지 확인
-			f.delete();
+		if (ntatVo != null) {
+			String fileName = ntatVo.getNtatOrigin();
+			File f =  new File(savePath +  fileName);
+			if(f.exists()) { //파일 존재하는지 확인
+				f.delete();
+			}
 		}
 		
 		//파일 유무 확인
 		if (ntvo.getNtFileName() != null && !ntvo.getNtFileName().isEmpty()) {
+			
+			
 			//파일 있음
 			//파일 업로드 후 저장된 파일명 얻기 
 			String changeName = FileUploader.fileUpload(ntvo.getNtFileName(), savePath);
